Move more BGP types from `omicron-common` to `nexus-types` (#9921)
Moves the latest versions of `BgpConfig`, `BgpImported`, and
`BgpExported` to the nexus-types version where they were introduced.
Allows us to address some lingering todos about `From` impls in the
wrong places on older versions.

Part of #9801.
