+++ common/src/api/external/mod.rs
@@ -3202,25 +3202,6 @@ pub struct SwitchPortRouteConfig {
     pub rib_priority: Option<u8>,
 }
 
-/// A base BGP configuration.
-#[derive(
-    ObjectIdentity, Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq,
-)]
-pub struct BgpConfig {
-    #[serde(flatten)]
-    pub identity: IdentityMetadata,
-
-    /// The autonomous system number of this BGP configuration.
-    pub asn: u32,
-
-    /// Optional virtual routing and forwarding identifier for this BGP
-    /// configuration.
-    pub vrf: Option<String>,
-
-    /// Maximum number of paths to use when multiple "best paths" exist
-    pub max_paths: MaxPathConfig,
-}
-
 /// An IP address configuration for a port settings object.
 #[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
 pub struct SwitchPortAddressConfig {
@@ -3263,35 +3244,6 @@ pub struct SwitchPortAddressView {
 
     /// The interface name this address belongs to.
     pub interface_name: Name,
-}
-
-/// Route exported to a peer.
-#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
-pub struct BgpExported {
-    /// Identifier for the BGP peer.
-    pub peer_id: String,
-
-    /// Switch the route is exported from.
-    pub switch: SwitchLocation,
-
-    /// The destination network prefix.
-    pub prefix: oxnet::IpNet,
-}
-
-/// A route imported from a BGP peer.
-#[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, PartialEq)]
-pub struct BgpImported {
-    /// The destination network prefix.
-    pub prefix: oxnet::IpNet,
-
-    /// The nexthop the prefix is reachable through.
-    pub nexthop: IpAddr,
-
-    /// BGP identifier of the originating router.
-    pub id: u32,
-
-    /// Switch the route is imported into.
-    pub switch: SwitchLocation,
 }
 
 /// BFD connection mode.
