+++ nexus/src/app/bgp.rs
@@ -9,8 +9,7 @@ use nexus_db_queries::context::OpContext;
 use nexus_types::external_api::networking;
 use omicron_common::api::external::http_pagination::PaginatedBy;
 use omicron_common::api::external::{
-    self, BgpExported, BgpImported, CreateResult, DeleteResult, ListResultVec,
-    LookupResult, NameOrId,
+    self, CreateResult, DeleteResult, ListResultVec, LookupResult, NameOrId,
 };
 
 impl super::Nexus {
@@ -152,7 +151,7 @@ impl super::Nexus {
     pub async fn bgp_exported(
         &self,
         opctx: &OpContext,
-    ) -> LookupResult<Vec<BgpExported>> {
+    ) -> LookupResult<Vec<networking::BgpExported>> {
         opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
         let mut result = vec![];
         for (switch, client) in &self.mg_clients().await.map_err(|e| {
@@ -204,7 +203,7 @@ impl super::Nexus {
                                 ))
                             }
                         };
-                        let export = BgpExported {
+                        let export = networking::BgpExported {
                             peer_id: peer_id.clone(),
                             switch: *switch,
                             prefix,
@@ -264,7 +263,7 @@ impl super::Nexus {
         &self,
         opctx: &OpContext,
         _sel: &networking::BgpRouteSelector,
-    ) -> ListResultVec<BgpImported> {
+    ) -> ListResultVec<networking::BgpImported> {
         opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
         let mut result = Vec::new();
         for (switch, client) in &self.mg_clients().await.map_err(|e| {
@@ -272,7 +271,7 @@ impl super::Nexus {
                 "failed to get mg clients: {e}"
             ))
         })? {
-            let mut imported: Vec<BgpImported> = Vec::new();
+            let mut imported: Vec<networking::BgpImported> = Vec::new();
             match client.get_rib_imported_v2(None, None).await {
                 Ok(result) => {
                     for (prefix, paths) in result.into_inner().iter() {
@@ -287,7 +286,7 @@ impl super::Nexus {
                             }
                         };
                         for p in paths.iter() {
-                            let x = BgpImported {
+                            let x = networking::BgpImported {
                                 switch: *switch,
                                 prefix: ipnet,
                                 id: p
