+++ nexus/src/external_api/http_entrypoints.rs
@@ -77,9 +77,6 @@ use omicron_common::api::external::AddressLotCreateResponse;
 use omicron_common::api::external::AddressLotViewResponse;
 use omicron_common::api::external::AffinityGroupMember;
 use omicron_common::api::external::AntiAffinityGroupMember;
-use omicron_common::api::external::BgpConfig;
-use omicron_common::api::external::BgpExported;
-use omicron_common::api::external::BgpImported;
 use omicron_common::api::external::DataPageParams;
 use omicron_common::api::external::Disk;
 use omicron_common::api::external::Error;
@@ -4322,11 +4319,11 @@ impl NexusExternalApi for NexusExternalApiImpl {
     async fn networking_bgp_config_create(
         rqctx: RequestContext<ApiContext>,
         config: TypedBody<networking::BgpConfigCreate>,
-    ) -> Result<HttpResponseCreated<BgpConfig>, HttpError> {
+    ) -> Result<HttpResponseCreated<networking::BgpConfig>, HttpError> {
         audit_and_time(&rqctx, |opctx, nexus| async move {
             let config = config.into_inner();
             let result = nexus.bgp_config_create(&opctx, &config).await?;
-            Ok(HttpResponseCreated::<BgpConfig>(result.try_into()?))
+            Ok(HttpResponseCreated::<networking::BgpConfig>(result.try_into()?))
         })
         .await
     }
@@ -4334,7 +4331,8 @@ impl NexusExternalApi for NexusExternalApiImpl {
     async fn networking_bgp_config_list(
         rqctx: RequestContext<ApiContext>,
         query_params: Query<PaginatedByNameOrId>,
-    ) -> Result<HttpResponseOk<ResultsPage<BgpConfig>>, HttpError> {
+    ) -> Result<HttpResponseOk<ResultsPage<networking::BgpConfig>>, HttpError>
+    {
         let apictx = rqctx.context();
         let handler = async {
             let nexus = &apictx.context.nexus;
@@ -4384,7 +4382,7 @@ impl NexusExternalApi for NexusExternalApiImpl {
 
     async fn networking_bgp_exported(
         rqctx: RequestContext<ApiContext>,
-    ) -> Result<HttpResponseOk<Vec<BgpExported>>, HttpError> {
+    ) -> Result<HttpResponseOk<Vec<networking::BgpExported>>, HttpError> {
         let apictx = rqctx.context();
         let opctx = crate::context::op_context_for_external_api(&rqctx).await?;
         let handler = async {
@@ -4449,7 +4447,7 @@ impl NexusExternalApi for NexusExternalApiImpl {
     async fn networking_bgp_imported(
         rqctx: RequestContext<ApiContext>,
         query_params: Query<networking::BgpRouteSelector>,
-    ) -> Result<HttpResponseOk<Vec<BgpImported>>, HttpError> {
+    ) -> Result<HttpResponseOk<Vec<networking::BgpImported>>, HttpError> {
         let apictx = rqctx.context();
         let opctx = crate::context::op_context_for_external_api(&rqctx).await?;
         let handler = async {
